rename: view 폴더 하위로 community 폴더 이동
