+++ app/src/main/java/com/umc/one_person_households_platform/view/community/CommunicateFragment.kt
@@ -1,0 +1,28 @@
+package com.umc.one_person_households_platform.view.community
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.umc.one_person_households_platform.R
+
+
+class CommunicateFragment : Fragment() {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_communicate, container, false)
+    }
+
+
+}
