+++ app/src/main/res/layout/fragment_communicate.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".community.CommunicateFragment">
+    tools:context=".view.community.CommunicateFragment">
 
     <!-- TODO: Update blank fragment layout -->
 <!--    <androidx.recyclerview.widget.RecyclerView-->
