+++ app/src/main/res/layout/fragment_newpost.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".community.NewpostFragment"
+    tools:context=".view.community.NewpostFragment"
     android:orientation="vertical">
 
     <!-- TODO: Update blank fragment layout -->
