+++ app/src/main/res/layout/fragment_postdetail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".community.PostdetailFragment">
+    tools:context=".view.community.PostdetailFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/my_toolbar"
