+++ app/src/main/res/layout/fragment_report.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".community.ReportFragment"
+    tools:context=".view.community.ReportFragment"
     android:orientation="vertical">
 
     <!-- TODO: Update blank fragment layout -->
