+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".community.SearchFragment"
+    tools:context=".view.community.SearchFragment"
     android:orientation="vertical">
 
     <androidx.appcompat.widget.Toolbar
